+++ regression/reference/strings-cfi/wrapstrings.cpp
@@ -1044,7 +1044,7 @@ void STR_fetchArrayStringArg_CFI(CFI_cdesc_t *SHT_strs_cfi)
     std::string *strs;
     int nstrs;
     fetchArrayStringArg(&strs, &nstrs);
-    // Copy results into strs
+    STR_ShroudCFIArrayStringOut(SHT_strs_cfi, strs, nstrs);
     // splicer end function.fetchArrayStringArg_CFI
 }
 
@@ -1098,7 +1098,20 @@ void STR_fetchArrayStringAlloc_CFI(CFI_cdesc_t *SHT_strs_cfi)
     std::string *strs;
     int nstrs;
     fetchArrayStringAlloc(&strs, &nstrs);
-    // Allocate and copy into strs
+    size_t SHT_strs_len = 0;
+    if (SHT_strs_len <= 0) {
+        SHT_strs_len = STR_ShroudArrayStringOutSize(strs, nstrs);
+    }
+    if (strs != nullptr) {
+        CFI_index_t SHT_strs_lower[1] = {1};
+        CFI_index_t SHT_strs_extents[1];
+        SHT_strs_extents[0] = nstrs;
+        int SH_ret = CFI_allocate(SHT_strs_cfi, SHT_strs_lower, 
+            SHT_strs_extents, SHT_strs_len);
+        if (SH_ret == CFI_SUCCESS) {
+            STR_ShroudCFIArrayStringOut(SHT_strs_cfi, strs, nstrs);
+        }
+    }
     // splicer end function.fetchArrayStringAlloc_CFI
 }
 
@@ -1152,7 +1165,20 @@ void STR_fetchArrayStringAllocLen_CFI(CFI_cdesc_t *SHT_strs_cfi)
     std::string *strs;
     int nstrs;
     fetchArrayStringAllocLen(&strs, &nstrs);
-    // Allocate and copy into strs
+    size_t SHT_strs_len = 20;
+    if (SHT_strs_len <= 0) {
+        SHT_strs_len = STR_ShroudArrayStringOutSize(strs, nstrs);
+    }
+    if (strs != nullptr) {
+        CFI_index_t SHT_strs_lower[1] = {1};
+        CFI_index_t SHT_strs_extents[1];
+        SHT_strs_extents[0] = nstrs;
+        int SH_ret = CFI_allocate(SHT_strs_cfi, SHT_strs_lower, 
+            SHT_strs_extents, SHT_strs_len);
+        if (SH_ret == CFI_SUCCESS) {
+            STR_ShroudCFIArrayStringOut(SHT_strs_cfi, strs, nstrs);
+        }
+    }
     // splicer end function.fetchArrayStringAllocLen_CFI
 }