+++ shroud/fcfmt.py
@@ -731,9 +731,7 @@ class FillFormat(object):
             # A string becomes an array of CHARACTER.
             fmt.i_dimension = "(*)"
 
-        if "len" in f_attrs:
-            fmt.f_char_len = "len=%s" % f_attrs["len"];
-        elif hasattr(fmt, "f_var_cdesc"):
+        if hasattr(fmt, "f_var_cdesc"):
             if meta["deref"] == "allocatable":
                 # Use elem_len from the C wrapper.
                 fmt.f_char_type = wformat("character(len={f_var_cdesc}%elem_len) ::\t ", fmt)